refactor: extract settings cluster overlays (#357)
Eighth slice. Bundles four settings-related overlays:

- draw_settings                  (~145 lines) - main settings overlay with grouped toggles
- draw_customize                 (~22 lines)  - Theme/Keybindings/Profile launcher
- draw_settings_profile_manager  (~100 lines) - profile manager with builtin/user contextual hints
- draw_settings_profile_save_as  (~30 lines)  - private helper of profile_manager, kept in same file

SETTINGS, SettingDef, and the SETTINGS_SECTION_* constants are now only
imported by ui/overlays/settings.rs, so the import block in ui/mod.rs
shed three more entries.

ui/mod.rs: 4434 -> 4132 lines.
